+++ NEWS.md
@@ -1,4 +1,5 @@
 # eatFDZ 0.7.0.9000
+* fixed `compare_data()` for cases with meta data differences either only on value or only variable level 
 
 # eatFDZ 0.7.0