+++ NEWS.md
@@ -1,5 +1,7 @@
 # eatFDZ 0.7.0.9000
 * fixed `compare_data()` for cases with meta data differences either only on value or only variable level 
+* `compare_data()` now uses `eatGADS::equalMeta()` instead of `eatGADS::equalGADS()`, reducing unnecessary overhead 
+* renamed `compare_data()` output (#30)
 
 # eatFDZ 0.7.0